refactor: defer to requesting/parsing image paths in the Gallery component
