+++ src/components/Gallery.tsx
@@ -1,4 +1,4 @@
-import ImgixAPI from 'imgix-management-js';
+import ImgixAPI, { APIError } from 'imgix-management-js';
 import { Component } from 'react';
 import { SourceProps } from './Dialog';
 
@@ -22,6 +22,43 @@ export default class Gallery extends Component<GalleryProps, GalleryState> {
     };
   }
 
+  getImages = async () => {
+    return await this.state.imgix.request(
+      `assets/${this.props.selectedSource?.id}`,
+    );
+  };
+
+  getImagePaths = async () => {
+    let images,
+      allOriginPaths: string[] = [];
+
+    try {
+      images = await this.getImages();
+    } catch (error) {
+      // APIError will emit more helpful data for debugging
+      if (error instanceof APIError) {
+        console.error(error.toString());
+      } else {
+        console.error(error);
+      }
+      return allOriginPaths;
+    }
+
+    /*
+     * Resolved requests can either return an array of objects or a single
+     * object via the `data` top-level field. When parsing all enabled sources,
+     * both possibilities must be accounted for.
+     */
+    const imagesArray = Array.isArray(images.data)
+      ? images.data
+      : [images.data];
+    imagesArray.map((image: any) =>
+      // TODO: add more explicit types for image
+      allOriginPaths.push(image.attributes.origin_path),
+    );
+
+    return allOriginPaths;
+  };
 
   render() {
     return (
